Clean up treatment of inhomogeneous Dirichlet BCs (#125)
Clean up class `DiscreteEquation`:

* Move class to new module `psydac.api.equation`;
* Add function `l2_boundary_projection` which creates an auxiliary equation for treating inhomogeneous Dirichlet BCs through the L2 projection of boundary data to the trial space (earlier this was handled in method `solve`); `l2_boundary_projection`returns an object of type `sympde.expr.Equation`, or `None` if not needed;
* `__init__` constructor creates L2 boundary projection, discretizes it and stores it in read-only property `self.boundary_equation`;
* `solve` method calls `self.assembly()` and, if needed, `self.boundary_equation.solve()`;
* `assembly` method reassembles the linear system only if free arguments are present, and does not accept the old keyword arguments `assemble_lhs` and `assemble_rhs` because their default values are unclear.
